% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean.R
\name{.get_mean}
\alias{.get_mean}
\title{Sequence of Variable Means}
\usage{
.get_mean(data, ng = 1001)
}
\arguments{
\item{data}{- dataframe or matrix containing all variables in the model (must be all numeric)}

\item{ng}{- optional length of the sequence}
}
\value{
- matrix of mean values
}
\description{
Sequence of Variable Means
}
\examples{
#Set length of sequence
n = 1001

#Load data
data(mtcars)
#Show number of parameters in data
ncol(mtcars)

#Apply get_means to dataset
z = .get_mean(mtcars, n)
#Show output
z
#Dimensions is n x (number of parameters)
dim(z)
}
