% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pred.R
\name{.get_pred}
\alias{.get_pred}
\title{Create Predicted Values + Upper and Lower Bounds for Confidence Interval}
\usage{
.get_pred(fit, new_data)
}
\arguments{
\item{fit}{- model used for making predictions}

\item{new_data}{- dataframe or matrix used for prediction}
}
\value{
- list of predictions and lower and upper bounds of confidence interval
}
\description{
Create Predicted Values + Upper and Lower Bounds for Confidence Interval
}
\examples{
#Load data
data(mtcars)

#Create simple regression model
fit = stats::lm(mpg ~ wt, data = mtcars)

#Create new set of data for prediction
new_wt = seq(min(mtcars$wt), max(mtcars$wt))
new_data = data.frame(wt = new_wt)

#Use get_pred to get predictions and confidence bounds
mpg_pred = get_pred(fit, new_data)

#Show output for predicted y
mpg_pred$fHat

#Show output for upper and lower bounds
mpg_pred$ub
mpg_pred$lb
}
