% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean.R
\name{get_mean}
\alias{get_mean}
\title{Sequence of Variable Means}
\usage{
get_mean(data, ng)
}
\arguments{
\item{data}{- dataframe or matrix containing all variables in the model}

\item{ng}{- length of the sequence}
}
\description{
Sequence of Variable Means
}
