% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Create call to model argument}
\usage{
get_model(y, smooth_terms, linear_terms)
}
\arguments{
\item{y}{- a string of the name of the predictor variable}

\item{smooth_terms}{- a vector of strings of the names of the spline terms}

\item{linear_terms}{- a vector of strings of the names of the linear terms}
}
\description{
Create call to model argument
}
\examples{
y = "house_price"
smooth_terms = c("sq_ft")
linear_terms = c("num_bedrooms", "num_bathrooms")

get_model(y, smooth_terms, linear_terms)
}
