% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot.R
\name{get_plot}
\alias{get_plot}
\title{Title}
\usage{
get_plot(x, fHat, yrange, title, x_lab, y_lab)
}
\arguments{
\item{x}{- dataframe or matrix containing spline terms}

\item{fHat}{- matrix of prediction values}

\item{yrange}{- matrix containing minimum and maximum for y axis}

\item{title}{- string for title of the plot}

\item{x_lab}{- string for title of x axis}

\item{y_lab}{- string for title of y axis}

\item{color}{- color of the line on the plot}
}
\description{
Title
}
