% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pred.R
\name{get_pred}
\alias{get_pred}
\title{Title}
\usage{
get_pred(new_data, type)
}
\arguments{
\item{new_data}{- dataframe or matrix used for prediction}

\item{type}{- string that gives type of prediction (either "response" or "link")}
}
\description{
Title
}
