% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_X_seq.R
\name{get_X_seq}
\alias{get_X_seq}
\title{Sequence of Spline Terms}
\usage{
get_X_seq(data, ng)
}
\arguments{
\item{data}{- matrix or dataframe containing only spline terms}

\item{ng}{- length of sequence}
}
\description{
Sequence of Spline Terms
}
