% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scamplot.R
\name{make_scamplot}
\alias{make_scamplot}
\title{Create Prediction Plot of SCAM}
\usage{
make_scamplot(
  data,
  y,
  smooth_terms,
  linear_terms,
  shape_type,
  type,
  title = "Prediction Plot"
)
}
\arguments{
\item{data}{- data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model}

\item{y}{- string containing name of depednent variable in model}

\item{smooth_terms}{- vector of strings containing names of independent variables to be fit as splines in the model}

\item{linear_terms}{- vector of strings containing names of independent variables to be fit as linear terms in the model}

\item{shape_type}{- vector of strings containing the shape constraints for the spline (in the same order as smooth_terms). Can only contain shape constraints that the scam function supports}

\item{type}{- string indicating "link" or "response"}

\item{title}{- optional string containing title of plot}
}
\value{
None
}
\description{
Create Prediction Plot of SCAM
}
\examples{
#Load the packages required
library(HRW)

#Load the data
data(BostonMortgages)

#Assign variables to be supplied to function
y = "deny"
smooth_terms = c("dir", "lvr")
BostonMortgages$ccs = as.factor(BostonMortgages$ccs)
linear_terms = c("ccs", "black", "pbcr", "self", "single")
shape_type = c("cr", "mpi")

#Use package to build predictive plot
make_scamplot(BostonMortgages, y, smooth_terms, linear_terms, shape_type, type = "link")
make_scamplot(BostonMortgages, y, smooth_terms, linear_terms, shape_type, type = "response")
}
