% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Create Formula for Call to Model}
\usage{
get_model(y, smooth_terms, linear_terms, shape_type)
}
\arguments{
\item{y}{- a string of the name of the predictor variable}

\item{smooth_terms}{- a vector of strings of the names of the spline terms}

\item{linear_terms}{- a vector of strings of the names of the linear terms}

\item{shape_type}{- a vector of strings containing the shape constraints for the spline (in the same order as smooth_terms). Can only contain shape constraints that the scam function supports}
}
\value{
- string of call to model
}
\description{
Create Formula for Call to Model
}
\examples{
y = "house_price"
smooth_terms = c("sq_ft", "distance_from_work")
linear_terms = c("num_bedrooms", "num_bathrooms")
shape_type = c("mpi", "mpd")

get_model(y, smooth_terms, linear_terms, shape_type)
}
